# api.yaml
openapi: '3.0.3'
info:
  title: MIFARE card reader API
  version: '1.0'
servers:
  - url: http://localhost:3333/

paths:

  /WriteData:
    post:
      summary: Записать данные на карту
      description: Ищет и записывает данные на указанную карту
      parameters:
        - name: ParkingEnterTime
          in: query
          description: Время въезда на парковку
          schema:
            type: string
            format: date-time
            example: '2015-10-02 15:16:37'
        - name: LastRecountTime
          in: query
          description: Время пересчёта
          schema:
            type: string
            format: date-time
            example: '2015-10-02 15:16:37'
        - name: SumOnCard
          in: query
          description: Сумма денег на карте
          schema:
            type: integer
            example: 23
        - name: TSidFC
          in: query
          description: Тарифная сетка
          schema:
            type: integer
            example: 1
        - name: TPidFC
          in: query
          description: Тарифный план
          schema:
            type: integer
            example: 2
        - name: ZoneidFC
          in: query
          description: Парковочная зона
          schema:
            type: integer
            example: 3
        - name: ClientGroupidFC
          in: query
          description: Группа доступа
          schema:
            type: integer
            example: 4
        - name: LastPaymentTime
          in: query
          description: Время последней оплаты
          schema:
            type: string
            format: date-time
            example: '2015-10-02 15:16:37'
        - name: Nulltime1
          in: query
          description: Null time 1
          schema:
            type: string
            format: date-time
            example: '2015-10-02 15:16:37'
        - name: Nulltime2
          in: query
          description: Null time 2
          schema:
            type: string
            format: date-time
            example: '2015-10-02 15:16:37'
        - name: Nulltime3
          in: query
          description: Null time 3
          schema:
            type: string
            format: date-time
            example: '2015-10-02 15:16:37'
        - name: TVP
          in: query
          description: TVP
          schema:
            type: string
            format: date-time
            example: '2015-10-02 15:16:37'
        - name: ClientTypidFC
          in: query
          description: ClientTypidFC
          schema:
            type: integer
            example: 3
        - name: TKVP
          in: query
          description: TKVP
          schema:
            type: integer
            example: 3
        - name: CardId
          in: query
          description: Номер карты в Hex представление
          schema:
            type: string
            pattern: ^[0-9a-fA-F]{2,8}$
            example: AB4569
        - name: callback
          in: query
          required: false
          schema:
            type: string

      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  Status:
                    $ref: '#/components/schemas/Status'
                  Now:
                    type: string
                    format: date-time
                    example: '2015-10-02 14:01:49'
                  CurrentStatus:
                    $ref: '#/components/schemas/CurrentStatus'

  /GetData:
    get:
      summary: Поиск карты
      description: Поиск и чтение карты
      parameters:
        - name: Sector
          in: query
          description: Номер сектора от 1 до 15
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 15
            example: 1
        - name: CardKey
          in: query
          description: Код доступа к сектору
          required: true
          schema:
            type: string
            pattern: ^[a-fA-F0-9]{2}:[a-fA-F0-9]{2}:[a-fA-F0-9]{2}:[a-fA-F0-9]{2}:[a-fA-F0-9]{2}:[a-fA-F0-9]{2}$
            example: FF:FF:FF:AA:BB:12
        - name: callback
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Результат поиска
          content:
            application/json:
              schema:
                type: object
                properties:
                  Status:
                    $ref: '#/components/schemas/Status'
                  Now:
                    type: string
                    format: date-time
                    example: '2015-10-02 15:16:37'
                  CurrentStatus:
                    $ref: '#/components/schemas/CurrentStatus'
                    description: NotUsed
                  CardId:
                    description: Номер карты в Hex представление
                    type: string
                    pattern: ^[0-9a-fA-F]{2,8}$
                    example: 12ABCD59
                  ParkingEnterTime:
                    description: Время въезда на парковку
                    type: string
                    format: date-time
                    example: '2015-10-02 15:16:37'
                  LastRecountTime:
                    description: Время пересчёта
                    type: string
                    format: date-time
                    example: '2015-10-02 15:16:37'
                  SumOnCard:
                    description: Сумма денег на карте
                    type: integer
                    example: 54
                  TSidFC:
                    description: Тарифная сетка
                    type: integer
                    example: 8
                  TPidFC:
                    description: Тарифный план
                    type: integer
                    example: 9
                  ZoneidFC:
                    description: Парковочная зона
                    type: integer
                    example: 2
                  ClientGroupidFC:
                    description: Группа доступа
                    type: integer
                    example: 1
                  LastPaymentTime:
                    description: Время последней оплаты
                    type: string
                    format: date-time
                    example: '2015-10-02 15:16:37'
                  Nulltime1:
                    type: string
                    format: date-time
                    example: '2000-01-01 00:00:00'
                  Nulltime2:
                    type: string
                    format: date-time
                    example: '2000-01-01 00:00:00'
                  Nulltime3:
                    type: string
                    format: date-time
                    example: '2000-01-01 00:00:00'
                  TVP:
                    type: string
                    format: date-time
                    example: '2000-01-01 00:00:00'
                  TKVP:
                    type: integer
                    example: 0
                  Regular:
                    type: boolean
components:
  schemas:
    Status:
      type: string
      enum:
        - CardReaded
        - SomeoneElsesCard
      example: CardReaded

    CurrentStatus:
      type: string
      enum:
        - ReaderNotExists
        - Unknown
        - CardExists
        - Exception
        - CardIDExists
        - CardReaded
        - ReadError
        - CardNotExists
        - CardIDNotExists
        - SomeoneElsesCard
        - CardWrited
        - WriteError
      example: CardReaded